Simulating Multi-sample Multi-cell-type scRNA-seq Dataset based on Negative Binomial Distribution

simuRNAseq simulates scRNA-seq data with multiple subjects (samples), multiple clusters (cell-types) and two treatments (conditions) based on a negative binomial (NB) distribution using a reference data as background or control. The reference data consisting of genes-by-cells counts matrix is used to estimate the NB dispersion and means for the genes to be simulated. The simulated genes are randomly selected from the reference data. The NB dispersion are estimated by the method-of-moments estimate (MME). The NB means for the background in the control are estimated by the sample mean. The NB means for the differentially expressed (DE) genes are given by the sample mean plus a log-fold change (logFC). The simulated cells are randomly selected from the meta data that specifies subjects, cell-types and treatments for the cells. The meta data consists of samples, clusters of cell types, and treatments, which can be generated either from reference data or randomly. If not provided, it will be randomly generated. A random seed is recommended to be specified by set.seed before simulating data.

Usage

simuRNAseq(
  counts,
  nGenes = nrow(counts),
  nCells = ncol(counts),
  metadata = NULL,
  samples.nested = TRUE,
  nsam = 25,
  ncls = 10,
  ntrt = 2,
  trt = NULL,
  nDEgenes = ncls,
  nDEgenesType,
  pDEgenesType = NULL,
  adjust.library.size = TRUE,
  direction = c("both", "up", "down"),
  minbeta = 0.25,
  maxbeta = 1,
  var.randomeffects = 0.1
)

Arguments

counts

A genes-by-cells matrix of reference counts. If missing, counts is generated by a negative binomial distribution.

nGenes

Number of genes to be simulated.

nCells

Number of cells to be simulated.

metadata

The meta data consisting of 4 columns: sam (sample labels), cls (cluster lables of cell types), trt (treatments or conditions) and libsize (library size or total counts per cell), which is randomly generated if not provided.

samples.nested

If TRUE, when metadata is not provided, each simulated subject (sample) belongs to only one condition (either treatment or control), that is, the subject is nested in a condition (treatment).

nsam

Number of subjects (individuals).

ncls

Number of clusters (cell-types).

ntrt

Number of treatments (only one condition is treated).

trt

Treatment, specifying which condition is treated.

nDEgenes

Total number of DE genes.

nDEgenesType

Number of DE genes specific to a cell type, named by cell cluster labels.

pDEgenesType

Proportion of DE genes in a cell-type. Default NULL means equal proportion.

adjust.library.size

If TRUE, adjust library sizes using the reference counts.

direction

Specify if the DE genes are up- and/or down-regulated.

minbeta

Lower bound of the DE gene logFC.

maxbeta

Upper bound of the DE gene logFC. minbeta < maxbeta. If direction = "both", minbeta*maxbeta > 0. If direction = "up", minbeta > 0. If direction = "down", maxbeta < 0.

var.randomeffects

Variance of random effects

Value

ref.mean.dispersion: A data frame of the reference counts' means and dispersion,

metadata: Meta data consisting of 4 columns: sam (sample labels), cls (cluster lables of cell types), trt (two treatments/conditions) and libsize (library sizes).

counts: A genes-by-cells matrix of the simulated counts.

DEgenes: A data frame of DE genes consisting of 3 columns: gene, beta (effect), and cluster to which the gene is specific.

treatment: The condition treated.

Examples

#Simulate a multi-sample multi-cell-type scRNA-seq dataset.
set.seed(2412)
dat <- simuRNAseq(nGenes = 50, nCells = 1000, nsam = 25, ncls = 4, ntrt = 2, nDEgenes = 6)
#> Message: the condition B is treated.
str(dat)
#> List of 5
#>  $ ref.mean.dispersion:'data.frame':	50 obs. of  2 variables:
#>   ..$ mu        : num [1:50] 1.077 0.992 2.815 1.945 2.06 ...
#>   ..$ dispersion: num [1:50] 1.602 2.989 0.954 2.937 1.914 ...
#>  $ metadata           :'data.frame':	1000 obs. of  4 variables:
#>   ..$ sam    : chr [1:1000] "B1" "A6" "A2" "B8" ...
#>   ..$ cls    : chr [1:1000] "C4" "C3" "C1" "C1" ...
#>   ..$ trt    : chr [1:1000] "B" "A" "A" "B" ...
#>   ..$ libsize: num [1:1000] 117 75 101 80 123 113 125 95 77 95 ...
#>  $ counts             : num [1:50, 1:1000] 0 2 6 2 5 2 0 0 2 11 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:50] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>   .. ..$ : chr [1:1000] "Cell1" "Cell2" "Cell3" "Cell4" ...
#>  $ DEgenes            :'data.frame':	6 obs. of  3 variables:
#>   ..$ gene   : chr [1:6] "Gene45" "Gene46" "Gene47" "Gene48" ...
#>   ..$ beta   : num [1:6] 0.55 0.365 0.831 -0.948 -0.875 ...
#>   ..$ cluster: chr [1:6] "C1" "C2" "C3" "C3" ...
#>  $ treatment          : chr "B"
table(dat$metadata[, c("sam", "trt")]) #The samples are nested in a condition.
#>      trt
#> sam    A  B
#>   A1  33  0
#>   A10 40  0
#>   A11 36  0
#>   A12 29  0
#>   A13 39  0
#>   A14 35  0
#>   A2  38  0
#>   A3  42  0
#>   A4  33  0
#>   A5  32  0
#>   A6  35  0
#>   A7  33  0
#>   A8  44  0
#>   A9  44  0
#>   B1   0 42
#>   B10  0 35
#>   B11  0 40
#>   B2   0 44
#>   B3   0 32
#>   B4   0 47
#>   B5   0 42
#>   B6   0 50
#>   B7   0 40
#>   B8   0 63
#>   B9   0 52

#Analyze differentially expressed (DE) genes specific to a cell-type using LMM.
Y <- log(dat$counts + 1) #expressions (log-transformed counts)
X <- model.matrix(~ 0 + log(libsize) + cls + cls:trt, data = dat$metadata)
Z <- model.matrix(~ 0 + sam, data = dat$metadata)
d <- ncol(Z)

#Fitting LMM using cell-level data.
fit <- lmmfit(Y, X, Z, d = d)

#Fitting LMM using summary-level data.
ss <- sslmm(X, Y, Z)
fitss <- lmm(summary.stats = ss, d = d)
identical(fit, fitss)
#> [1] TRUE

#Hypothesis tests
test <- lmmtest(fit)
head(test)
#>       log(libsize)_t   clsC1_t   clsC2_t   clsC3_t   clsC4_t clsC1:trtB_t
#> Gene1       3.556438 -2.415894 -2.595711 -2.557626 -2.446634    0.8697778
#> Gene2       6.135313 -4.869716 -4.991892 -5.024973 -5.006548    0.3659413
#> Gene3       3.609899 -2.312274 -2.246792 -2.148500 -2.220267    0.1515357
#> Gene4       4.123944 -2.670994 -2.510841 -2.585615 -2.698308    0.9707563
#> Gene5       2.818000 -1.322021 -1.238786 -1.253600 -1.205876   -0.9577914
#> Gene6       3.592507 -2.552389 -2.766682 -2.699148 -2.635315    0.5414995
#>       clsC2:trtB_t clsC3:trtB_t clsC4:trtB_t log(libsize)_p      clsC1_p
#> Gene1    2.0700456    1.5066385   -0.3263183   3.936946e-04 1.587661e-02
#> Gene2    1.0976568    1.5001771    1.6810047   1.226867e-09 1.300111e-06
#> Gene3   -0.1112106   -0.3659261   -0.4559326   3.216502e-04 2.096680e-02
#> Gene4    0.7423556    0.8883383    0.6462646   4.036515e-05 7.686618e-03
#> Gene5   -0.5997369   -1.0410003   -1.7515738   4.928509e-03 1.864664e-01
#> Gene6    1.5196520    0.3102291    0.6687761   3.436079e-04 1.084772e-02
#>            clsC2_p      clsC3_p      clsC4_p clsC1:trtB_p clsC2:trtB_p
#> Gene1 9.579168e-03 1.068679e-02 1.459256e-02    0.3846325   0.03870667
#> Gene2 7.060831e-07 5.971329e-07 6.556356e-07    0.7144869   0.27262100
#> Gene3 2.487275e-02 3.191586e-02 2.662620e-02    0.8795840   0.91147190
#> Gene4 1.220264e-02 9.862323e-03 7.087769e-03    0.3319065   0.45804781
#> Gene5 2.157181e-01 2.102832e-01 2.281532e-01    0.3384015   0.54881868
#> Gene6 5.768368e-03 7.070027e-03 8.537187e-03    0.5882851   0.12891743
#>       clsC3:trtB_p clsC4:trtB_p
#> Gene1    0.1322220   0.74425245
#> Gene2    0.1338870   0.09307711
#> Gene3    0.7144983   0.64853836
#> Gene4    0.3745743   0.51825766
#> Gene5    0.2981293   0.08015644
#> Gene6    0.7564520   0.50379408

#The DE genes specific to a cell-type.
tail(test[, grep(":", colnames(test))])
#>        clsC1:trtB_t clsC2:trtB_t clsC3:trtB_t clsC4:trtB_t clsC1:trtB_p
#> Gene45    2.1720213   0.55889245   0.08637871   -0.9345875   0.03009058
#> Gene46    0.8149981   2.28175377  -0.73193986    0.8343811   0.41526921
#> Gene47    0.5897109   1.41986369   5.97111283    0.8070739   0.55551888
#> Gene48    0.2007912  -0.06210722  -4.68109321   -2.2274365   0.84090305
#> Gene49   -0.2895705   0.03168994  -3.22145147    0.9415348   0.77220544
#> Gene50    1.8874002   1.58847538   6.53482944    1.2079394   0.05939855
#>        clsC2:trtB_p clsC3:trtB_p clsC4:trtB_p
#> Gene45   0.57636139 9.311828e-01   0.35022860
#> Gene46   0.02271521 4.643783e-01   0.40426735
#> Gene47   0.15596186 3.279879e-09   0.41981754
#> Gene48   0.95048997 3.250047e-06   0.02614204
#> Gene49   0.97472569 1.316972e-03   0.34666032
#> Gene50   0.11249785 1.017003e-10   0.22735878
#The true DE genes
dat$DEgenes
#>      gene       beta cluster
#> 45 Gene45  0.5496463      C1
#> 46 Gene46  0.3652007      C2
#> 47 Gene47  0.8306937      C3
#> 48 Gene48 -0.9476624      C3
#> 49 Gene49 -0.8751216      C3
#> 50 Gene50  0.9050662      C3