Computing Summary-level Data from Individual-level Data

sslmm can be used to compute the correlation-related summary statistics (summary-level data) for lmm function.

Usage

sslmm(X, Y, Z)

Arguments

X

A design matrix for fixed effects, with rows corresponding to the columns of Y.

Y

A features-by-samples matrix of responses (genes-by-cells matrix of gene expressions for scRNA-seq).

Z

A design matrix for random effects, with rows corresponding to the columns of Y.

Value

A list of summary statistics: XX = t(X)%*%X, XY = t(X)%*%t(Y), ZX = t(Z)%*%X, ZY = t(Z)%*%t(Y), ZZ = t(Z)%*%Z, Ynorm = rowSums(Y*Y) and n = nrow(X).

Examples

n <- 1e3
set.seed(2024)
p <- 2
X <- matrix(rnorm(p*n), n, p)
colnames(X) <- paste0("X", 1:ncol(X))
m <- 3
Y <- matrix(rnorm(m*n), m, n)
rownames(Y) <- paste0("Y", 1:nrow(Y))
q <- 4
Z <- gl(q, n/q, labels = letters[1:q])
Z <- model.matrix(~ 0 + Z)
sslmm(X, Y, Z)
#> $n
#> [1] 1000
#> 
#> $XX
#>           X1        X2
#> X1 964.78381 -13.62653
#> X2 -13.62653 945.83427
#> 
#> $XY
#>           Y1        Y2         Y3
#> X1  34.59866 -58.55644 -30.742385
#> X2 -27.22278 -21.79678  -9.024452
#> 
#> $ZX
#>            X1        X2
#> Za  10.129790 -1.989424
#> Zb  15.517769 22.103361
#> Zc -16.168713  2.409996
#> Zd  -4.936147  3.201936
#> 
#> $ZY
#>             Y1          Y2        Y3
#> Za   6.2979384  -0.2567696 16.103129
#> Zb   0.8901537 -15.1695035 -5.891425
#> Zc -16.3589237  10.2965567 -4.849434
#> Zd  -6.7887706  -3.4119874  2.865243
#> 
#> $ZZ
#>     Za  Zb  Zc  Zd
#> Za 250   0   0   0
#> Zb   0 250   0   0
#> Zc   0   0 250   0
#> Zd   0   0   0 250
#> 
#> $Ynorm
#>        Y1        Y2        Y3 
#>  988.7703  991.0990 1019.9191 
#>