Fitting Linear Mixed-effects Models

lmmfit, a wrapper function of lmm, fits linear mixed-effects models (LMM) by sample-level data. The LMM parameters are estimated by restricted maximum likelihood (REML) with Fisher scoring (FS) gradient descent algorithm.

Usage

lmmfit(
  Y,
  X,
  Z,
  d,
  theta0 = NULL,
  method = "REML-FS",
  max.iter = 50,
  epsilon = 1e-05,
  output.cov = TRUE,
  output.RE = FALSE
)

Arguments

Y

A features-by-samples matrix of responses (genes-by-cells matrix of gene expressions for scRNA-seq).

X

A design matrix for fixed effects, with rows corresponding to the columns of Y.

Z

A design matrix for random effects, with rows corresponding to the columns of Y. Z = [Z1, ..., Zk], and Zi, i=1,...,k, is the design matrix for the i-th type random factor.

d

A vector of (m1,...,mk), mi = ncol(Zi), number of columns in Zi. m1 + ... + mk = ncol(Z), number of columns in Z.

theta0

A vector of initial values of the variance components, (s1, ...,sk, s_(k+1)), si = sigma_i^2, the variance component of the i-th type random effects. s_(k+1) = sigma^2, the variance component of model residual error.

method

The REML with Fisher scoring (FS) iterative algorithm, REML-FS.

max.iter

The maximal number of iterations for the iterative algorithm.

epsilon

Positive convergence tolerance. If the absolute value of the first partial derivative of log likelihood is less than epsilon, the iterations converge.

output.cov

If TRUE, output the covariance matrices for the estimated coefficients, which are needed for testing contrasts.

output.RE

If TRUE, output the best linear unbiased prediction (BLUP) of the random effects.

Value

A list containing the following components:

dlogL First partial derivatives of log-likelihoods for each feature (gene).

niter Nmbers of iterations for each feature (gene).

coef A matrix of estimated coefficients (fixed effects), each column corresponds to a feature (gene) and each row one covariate.

se A matrix of the standard errors of the estimated coefficients.

t A matrix of t-values for the fixed effects, equal to coef/se.

df Degrees of freedom.

p A matrix of two-sided p-values for the fixed effects.

cov A array of covariance matrices of the estimated coefficients (fixed effects).

theta A matrix of the estimated variance components, each column corresponds to a feature (gene) and each row one variance component. The last row is the variance component of the residual error.

se.theta Standard errors of the estimated theta.

RE A matrix of the best linear unbiased prediction (BLUP) of random effects.

Examples

#Generate data: X, Y, and Z.
set.seed(2024)

n <- 1e3
m <- 10
Y <- matrix(rnorm(m*n), m, n)
rownames(Y) <- paste0("Gene", 1:nrow(Y))

trt <- sample(c("A", "B"), n, replace = TRUE)
X <- model.matrix(~ 0 + trt)

q <- 20
sam <- rep(NA, n)
sam[trt == "A"] <- paste0("A", sample.int(q/2, sum(trt == "A"), replace = TRUE))
sam[trt == "B"] <- paste0("B", sample.int(q/2, sum(trt == "B"), replace = TRUE))
Z <- model.matrix(~ 0 + sam)
d <- ncol(Z)

#Fitting LMM by the cell-level data
fit <- lmmfit(Y, X, Z, d = d)
str(fit)
#> List of 12
#>  $ method  : chr "REML-FS"
#>  $ dlogL   : num [1:2, 1:10] -1.34e-06 -8.32e-09 -1.02e-06 -1.38e-09 9.52e-07 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:10] "dl" "dl" "dl" "dl" ...
#>  $ niter   : num [1:10] 10 7 6 8 7 7 5 6 10 11
#>  $ coef    : num [1:2, 1:10] 0.03745 0.01666 -0.04225 -0.02294 -0.00892 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "trtA" "trtB"
#>   .. ..$ : chr [1:10] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>  $ se      : num [1:2, 1:10] 0.0378 0.0376 0.0426 0.0424 0.0552 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "trtA" "trtB"
#>   .. ..$ : chr [1:10] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>  $ t       : num [1:2, 1:10] 0.992 0.443 -0.992 -0.541 -0.162 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "trtA" "trtB"
#>   .. ..$ : chr [1:10] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>  $ p       : num [1:2, 1:10] 0.322 0.658 0.321 0.589 0.872 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "trtA" "trtB"
#>   .. ..$ : chr [1:10] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>  $ cov     : num [1:2, 1:2, 1:10] 0.00143 0 0 0.00141 0.00181 ...
#>   ..- attr(*, "dimnames")=List of 3
#>   .. ..$ : chr [1:2] "trtA" "trtB"
#>   .. ..$ : chr [1:2] "trtA" "trtB"
#>   .. ..$ : chr [1:10] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>  $ df      : int 998
#>  $ theta   : num [1:2, 1:10] -0.00641 1.03645 -0.00132 0.97027 0.01133 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "var1" "var0"
#>   .. ..$ : chr [1:10] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>  $ se.theta: num [1:2, 1:10] 0.00478 0.04681 0.00606 0.04383 0.01013 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "var1" "var0"
#>   .. ..$ : chr [1:10] "Gene1" "Gene2" "Gene3" "Gene4" ...
#>  $ RE      : NULL

#Hypothesis tests
lmmtest(fit)
#>             trtA_t     trtB_t     trtA_p    trtB_p
#> Gene1   0.99151700  0.4430047 0.32167353 0.6578584
#> Gene2  -0.99188018 -0.5407880 0.32149640 0.5887744
#> Gene3  -0.16176056 -0.1851246 0.87152715 0.8531689
#> Gene4  -0.56751196 -0.8348433 0.57049404 0.4040058
#> Gene5  -0.97591263 -0.2341318 0.32934425 0.8149307
#> Gene6  -0.05543855 -0.7219293 0.95580018 0.4705070
#> Gene7  -1.17149564 -0.6587799 0.24167922 0.5101891
#> Gene8   0.10299581  2.2227291 0.91798699 0.0264575
#> Gene9   1.32183317  0.3175327 0.18652661 0.7509059
#> Gene10 -2.00516281  0.2376670 0.04521529 0.8121882
lmmtest(fit, index = 2)
#>            trtB_t    trtB_p
#> Gene1   0.4430047 0.6578584
#> Gene2  -0.5407880 0.5887744
#> Gene3  -0.1851246 0.8531689
#> Gene4  -0.8348433 0.4040058
#> Gene5  -0.2341318 0.8149307
#> Gene6  -0.7219293 0.4705070
#> Gene7  -0.6587799 0.5101891
#> Gene8   2.2227291 0.0264575
#> Gene9   0.3175327 0.7509059
#> Gene10  0.2376670 0.8121882
lmmtest(fit, contrast = cbind("B-A" = c(-1, 1)))
#>              B-A_t     B-A_p
#> Gene1  -0.39018951 0.6964797
#> Gene2   0.32121982 0.7481111
#> Gene3  -0.01616384 0.9871069
#> Gene4  -0.18729690 0.8514659
#> Gene5   0.52649028 0.5986646
#> Gene6  -0.47033528 0.6382183
#> Gene7   0.36502924 0.7151669
#> Gene8   1.49642518 0.1348591
#> Gene9  -0.71274410 0.4761707
#> Gene10  1.58872838 0.1124385